Add config and env modules to keep test config/environment in a single place.
The unittest modules and the test environment creation/deletion scripts 
currently use hardcoded queue manager and queue names and several environment
variables. For easier customization, these can be taken from the new config 
module. The env module is just for adding the checkout path to sys.path, which
currently is repeated in several (but not all) modules.