Make test usable with both lxml + minidom, use env.py for sys.path setting.
Fixes the dependency of this test on the full XML parser error text, which is
different for lxml and minidom. Previously, this test would only properly work
with lxml but not with minidom.
This test module can now be run standalone due to importing env.py which sets
sys.path properly.